Source: Restructurize code
We cannot have everything in the same 1 big struct that we pass to
basically every function. While it works for some code bases
it wont work here due to the absolute need for every little tiny
bit of CPU speed. Also we need board_t because of input thread plan